requirements.txt
# ===== Web Framework =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# ===== Configuration =====
python-dotenv==1.0.0
python-decouple==3.8
pydantic==2.5.0

# ===== HTTP Clients =====
httpx==0.25.2
aiohttp==3.9.5
aiofiles==23.2.1

# ===== OpenAI (включает Whisper API) =====
openai==1.3.7

# ===== Audio Processing =====
pydub==0.25.1

# ===== Redis & Caching =====
redis==5.0.1

# ===== PostgreSQL для долгосрочного хранения =====
sqlalchemy==2.0.23
psycopg2-binary==2.9.10

# ===== Task Queue (опционально) =====
celery==5.3.4

# ===== Logging =====
loguru==0.7.2

# ===== Security & Auth =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ===== Google APIs (для динамических промптов) =====
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1 